feat(settings): add DesktopShortcutsContent component for managing hotkeys and update user settings dialog
